+++ docs/release_notes.rst
@@ -25,6 +25,7 @@ What's New?
    data from :mod:`pudl`.
 *  New CLI built off a single command ``rmi`` with ``cloud`` and ``pudl`` subcommands
    for cleaning caches and configs.
+*  Added :func:`.simplify_strings` to :mod:`.pudl_helpers`.
 
 Bug Fixes
 ^^^^^^^^^